+++ src/test/suite/json/json.test.ts
@@ -1,0 +1,8 @@
+import { JsonStringParser } from "../../../providers/json";
+import { runJsonTests } from "./common";
+
+// ref: https://www.json.org/json-en.html
+
+suite("JSON", () => {
+  runJsonTests(new JsonStringParser());
+});