+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v0.1.8
+
+- Feat(json/jsonc/ts): error handling for invalid escape sequence with retsac v0.15.0 built-in string utils.
+- Perf(ts): apply `Action.prevent` and `Lexer.javascript.simpleStringLiteral` from retsac v0.15.0.
+
 ## v0.1.7
 
 - Feat(json/jsonc/ts): support unclosed string.